requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
google-generativeai==0.8.5
twilio==9.7.0
matplotlib==3.10.5
fpdf2==2.8.4
pytesseract==0.3.13
Pillow==11.3.0
supabase==2.17.0
streamlit==1.49.0
requests==2.32.5
